pyproject.toml
[project]
name = "pianist-mj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mjlab>=0.1.0",
    "mujoco-warp",
]

[tool.uv.sources]
mjlab = { path = "../mjlab" }
mujoco-warp = { git = "https://github.com/google-deepmind/mujoco_warp", rev = "e605c406363805d949b75d50dd5ffb3d79bc3f70" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pianist_mj"]
